Use SDL_UpdateTexture instead of SDL_LockTexture.
Note: The SDL3 documentation states that SDL_UpdateTexture is a fairly
slow function, intended for use with static textures that do not change
often, and that If the texture is intended to be updated often, it is
preferred to create the texture as streaming and use the locking
functions. In our case however the opposite is true.